+++ src/plonemeeting/portal/core/rest/site.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+import os
+
 from plone import api
 from plone.memoize import ram
 from plone.protect.interfaces import IDisableCSRFProtection
@@ -18,6 +20,7 @@ from zope.schema.interfaces import IVocabularyFactory
 
 import json
 import requests
+from datetime import datetime, timedelta
 
 
 class InstitutionLocationsAPIView(PublicAPIView):
@@ -82,3 +85,60 @@ class InstitutionLocationsAPIView(PublicAPIView):
 
         self.request.response.setHeader("Content-type", "application/json")
         return json.dumps(institution_locations)
+
+class PrometheusExporter(PublicAPIView):
+    """Publications Prometheus exporter"""
+
+    def __call__(self):
+        return self.publications_stats()
+
+    def publications_stats(self):
+        """
+        Get publications stats in Prometheus format
+        """
+        catalog = api.portal.get_tool(name="portal_catalog")
+        publications_published = catalog.unrestrictedSearchResults(portal_type="Publication", review_state="published")
+        res = "# TYPE publications_published gauge\n"
+        res += "# HELP publications_published Number of publications published\n"
+        res += "publications_published " + str(len(publications_published)) + "\n\n"
+
+        publications_planned = catalog.unrestrictedSearchResults(portal_type="Publication", review_state="planned")
+        res += "# TYPE publications_planned gauge\n"
+        res += "# HELP publications_planned Number of publications planned\n"
+        res += "publications_planned " + str(len(publications_planned)) + "\n\n"
+
+        now = datetime.now()
+        publications_planned_not_published = catalog.unrestrictedSearchResults(portal_type="Publication", review_state="planned", effective=now - timedelta(minutes=60))
+        res += "# TYPE publications_planned_late gauge\n"
+        res += "# HELP publications_planned_late Number of publications planned but not published\n"
+        res += "publications_planned_late " + str(len(publications_planned_not_published)) + "\n\n"
+
+        publications_expired = catalog.unrestrictedSearchResults(portal_type="Publication", review_state="published", expires=now)
+        res += "# TYPE publications_expired gauge\n"
+        res += "# HELP publications_expired Number of publications expired\n"
+        res += "publications_expired " + str(len(publications_expired)) + "\n\n"
+
+        publications_expired_not_removed = catalog.unrestrictedSearchResults(portal_type="Publication", review_state="published", expires=now - timedelta(minutes=60))
+        res += "# TYPE publications_expired_late gauge\n"
+        res += "# HELP publications_expired_late Number of publications expired but not removed\n"
+        res += "publications_expired_late " + str(len(publications_expired_not_removed)) + "\n\n"
+
+        dangling_publications = publications_planned_not_published + publications_expired_not_removed
+        res += "# TYPE dangling_publications gauge\n"
+        res += "# HELP dangling_publications Number of dangling publications\n"
+        res += "dangling_publications " + str(len(dangling_publications)) + "\n\n"
+
+        cron_log_path = "./var/log/cron.log"
+        if os.path.exists(cron_log_path):
+            last_cron_run = os.path.getmtime(cron_log_path)
+            res += "# TYPE last_cron_run gauge\n"
+            res += "# HELP last_cron_run Last cron run\n"
+            res += "last_cron_run " + str(last_cron_run) + "\n\n"
+
+            cron_is_not_running = last_cron_run < now.timestamp() - 3600
+            res += "# TYPE cron_is_not_running gauge\n"
+            res += "# HELP cron_is_not_running Cron is not running\n"
+            res += "cron_is_not_running " + str(cron_is_not_running) + "\n\n"
+
+        self.request.response.setHeader("Content-type", "text/plain")
+        return res